+++ tests/e2e/core-tests/CHANGELOG.md
@@ -20,9 +20,11 @@
 - Merchant Product Search tests
 - Shopper Single Product tests
 - Shopper Checkout Apply Coupon
+- Shopper Shop Browse Search Sort
 - Merchant Orders Customer Checkout Page
 - Shopper Cart Apply Coupon
 - Shopper Variable product info updates on different variations
+- Merchant order emails flow
 
 ## Fixed